+++ out/extension.js
@@ -10,18 +10,20 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
 Object.defineProperty(exports, "__esModule", { value: true });
 const vscode = require("vscode");
 const fs = require("fs");
+const fileSystem_1 = require("./fileSystem");
 const { copy } = require('copy-paste');
 const { basename } = require('path');
 const { exec } = require('child_process');
 function activate(context) {
     console.log('OWScript extension activated.');
-    const scheme = 'owscript';
-    const document = new OWScriptDocument();
-    context.subscriptions.push(vscode.workspace.registerTextDocumentContentProvider(scheme, document));
+    const config = vscode.workspace.getConfiguration('owscript');
+    const channel = vscode.window.createOutputChannel('OWScript');
+    let owfs = new fileSystem_1.OWScriptFS();
+    context.subscriptions.push(vscode.workspace.registerFileSystemProvider('owfs', owfs, { isCaseSensitive: true }));
     context.subscriptions.push(vscode.commands.registerTextEditorCommand('owscript.compile', (editor) => __awaiter(this, void 0, void 0, function* () {
+        channel.clear();
         const path = editor.document.uri.fsPath;
-        const config = vscode.workspace.getConfiguration('owscript');
-        const command = `cd ${config.path} & cat ${path} | python OWScript.py`;
+        const command = `set PYTHONIOENCODING=utf-8&cat "${path}" | python OWScript.py`;
         // Check if the OWScript directory exists using fs.access
         function pathExists(path) {
             return new Promise((res, rej) => {
@@ -30,22 +32,24 @@ function activate(context) {
                 }));
             });
         }
+        // Compile the source into workshop code
         function compileSource(command) {
             return new Promise((res, rej) => {
-                exec(command, (err, stdout, stderr) => __awaiter(this, void 0, void 0, function* () {
-                    err ? rej(err) : res({ stdout: stdout, stderr: stderr });
+                exec(command, {
+                    cwd: config.path
+                }, (err, stdout, stderr) => __awaiter(this, void 0, void 0, function* () {
+                    res({ stdout: stdout, stderr: stderr });
                 }));
             });
         }
         if (yield pathExists(config.path)) {
             let result = yield compileSource(command);
-            OWScriptDocument.output = result.stdout;
             if (result.stderr) {
                 let infoMessages = [];
                 let warningMessages = [];
                 let errorMessages = [];
                 for (let rawMsg of result.stderr.split('\n')) {
-                    let msg = rawMsg.replace(/^\[.*\]\s*/g, '');
+                    let msg = rawMsg.replace(/^\[[A-Z]*?\]\s*/g, '');
                     if (rawMsg.startsWith("[INFO]")) {
                         infoMessages.push(msg);
                     }
@@ -63,19 +67,22 @@ function activate(context) {
                     vscode.window.showWarningMessage(warningMessages.join('\n'));
                 }
                 if (errorMessages.length > 0) {
-                    vscode.window.showErrorMessage('Error while compiling ' + path + ' (see debug console)');
-                    console.error(errorMessages.join('\n'));
+                    vscode.window.showErrorMessage('Error while compiling ' + basename(path));
+                    channel.append(errorMessages.join('\n'));
+                    channel.show();
                     // Focus error automatically?
                 }
             }
             if (result.stdout != '') {
                 // Open the text document with the resulting compiled code
-                let uri = vscode.Uri.parse('owscript:' + basename(path) + ' - Workshop Code');
+                let code = Buffer.from(result.stdout);
+                let uri = vscode.Uri.parse('owfs:/' + basename(path) + ' - Workshop Code');
+                owfs.writeFile(uri, code, { create: true, overwrite: true });
                 let doc = yield vscode.workspace.openTextDocument(uri);
                 vscode.window.showTextDocument(doc, vscode.ViewColumn.Beside);
                 // Copy output
                 if (config.clipboard == true) {
-                    copy(OWScriptDocument.output, () => { });
+                    copy(result.stdout, () => { });
                     vscode.window.showInformationMessage('Code copied to clipboard.');
                 }
             }
@@ -87,17 +94,13 @@ function activate(context) {
             }
         }
     })));
+    vscode.workspace.onDidSaveTextDocument((doc) => __awaiter(this, void 0, void 0, function* () {
+        if (doc.languageId == 'owscript' && !doc.isUntitled && config.compileOnSave == true) {
+            yield vscode.commands.executeCommand('owscript.compile');
+        }
+    }));
 }
 exports.activate = activate;
 function deactivate() { }
 exports.deactivate = deactivate;
-class OWScriptDocument {
-    constructor() {
-        this.onDidChangeEmitter = new vscode.EventEmitter();
-        this.onDidChange = this.onDidChangeEmitter.event;
-    }
-    provideTextDocumentContent(uri) {
-        return OWScriptDocument.output;
-    }
-}
 //# sourceMappingURL=extension.js.map