+++ out/fileSystem.js
@@ -1,0 +1,167 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+const path = require("path");
+const vscode = require("vscode");
+class File {
+    constructor(name) {
+        this.type = vscode.FileType.File;
+        this.ctime = Date.now();
+        this.mtime = Date.now();
+        this.size = 0;
+        this.name = name;
+    }
+}
+exports.File = File;
+class Directory {
+    constructor(name) {
+        this.type = vscode.FileType.Directory;
+        this.ctime = Date.now();
+        this.mtime = Date.now();
+        this.size = 0;
+        this.name = name;
+        this.entries = new Map();
+    }
+}
+exports.Directory = Directory;
+class OWScriptFS {
+    constructor() {
+        this.root = new Directory('');
+        this._emitter = new vscode.EventEmitter();
+        this._bufferedEvents = [];
+        this.onDidChangeFile = this._emitter.event;
+    }
+    stat(uri) {
+        return this._lookup(uri, false);
+    }
+    readDirectory(uri) {
+        const entry = this._lookupAsDirectory(uri, false);
+        let result = [];
+        for (const [name, child] of entry.entries) {
+            result.push([name, child.type]);
+        }
+        return result;
+    }
+    // --- manage file contents
+    readFile(uri) {
+        const data = this._lookupAsFile(uri, false).data;
+        if (data) {
+            return data;
+        }
+        throw vscode.FileSystemError.FileNotFound();
+    }
+    writeFile(uri, content, options) {
+        let basename = path.posix.basename(uri.path);
+        let parent = this._lookupParentDirectory(uri);
+        let entry = parent.entries.get(basename);
+        if (entry instanceof Directory) {
+            throw vscode.FileSystemError.FileIsADirectory(uri);
+        }
+        if (!entry && !options.create) {
+            throw vscode.FileSystemError.FileNotFound(uri);
+        }
+        if (entry && options.create && !options.overwrite) {
+            throw vscode.FileSystemError.FileExists(uri);
+        }
+        if (!entry) {
+            entry = new File(basename);
+            parent.entries.set(basename, entry);
+            this._fireSoon({ type: vscode.FileChangeType.Created, uri });
+        }
+        entry.mtime = Date.now();
+        entry.size = content.byteLength;
+        entry.data = content;
+        this._fireSoon({ type: vscode.FileChangeType.Changed, uri });
+    }
+    rename(oldUri, newUri, options) {
+        if (!options.overwrite && this._lookup(newUri, true)) {
+            throw vscode.FileSystemError.FileExists(newUri);
+        }
+        let entry = this._lookup(oldUri, false);
+        let oldParent = this._lookupParentDirectory(oldUri);
+        let newParent = this._lookupParentDirectory(newUri);
+        let newName = path.posix.basename(newUri.path);
+        oldParent.entries.delete(entry.name);
+        entry.name = newName;
+        newParent.entries.set(newName, entry);
+        this._fireSoon({ type: vscode.FileChangeType.Deleted, uri: oldUri }, { type: vscode.FileChangeType.Created, uri: newUri });
+    }
+    delete(uri) {
+        let dirname = uri.with({ path: path.posix.dirname(uri.path) });
+        let basename = path.posix.basename(uri.path);
+        let parent = this._lookupAsDirectory(dirname, false);
+        if (!parent.entries.has(basename)) {
+            throw vscode.FileSystemError.FileNotFound(uri);
+        }
+        parent.entries.delete(basename);
+        parent.mtime = Date.now();
+        parent.size -= 1;
+        this._fireSoon({ type: vscode.FileChangeType.Changed, uri: dirname }, { uri, type: vscode.FileChangeType.Deleted });
+    }
+    createDirectory(uri) {
+        let basename = path.posix.basename(uri.path);
+        let dirname = uri.with({ path: path.posix.dirname(uri.path) });
+        let parent = this._lookupAsDirectory(dirname, false);
+        let entry = new Directory(basename);
+        parent.entries.set(entry.name, entry);
+        parent.mtime = Date.now();
+        parent.size += 1;
+        this._fireSoon({ type: vscode.FileChangeType.Changed, uri: dirname }, { type: vscode.FileChangeType.Created, uri });
+    }
+    _lookup(uri, silent) {
+        let parts = uri.path.split('/');
+        let entry = this.root;
+        for (const part of parts) {
+            if (!part) {
+                continue;
+            }
+            let child;
+            if (entry instanceof Directory) {
+                child = entry.entries.get(part);
+            }
+            if (!child) {
+                if (!silent) {
+                    throw vscode.FileSystemError.FileNotFound(uri);
+                }
+                else {
+                    return undefined;
+                }
+            }
+            entry = child;
+        }
+        return entry;
+    }
+    _lookupAsDirectory(uri, silent) {
+        let entry = this._lookup(uri, silent);
+        if (entry instanceof Directory) {
+            return entry;
+        }
+        throw vscode.FileSystemError.FileNotADirectory(uri);
+    }
+    _lookupAsFile(uri, silent) {
+        let entry = this._lookup(uri, silent);
+        if (entry instanceof File) {
+            return entry;
+        }
+        throw vscode.FileSystemError.FileIsADirectory(uri);
+    }
+    _lookupParentDirectory(uri) {
+        const dirname = uri.with({ path: path.posix.dirname(uri.path) });
+        return this._lookupAsDirectory(dirname, false);
+    }
+    watch(_resource) {
+        // ignore, fires for all changes...
+        return new vscode.Disposable(() => { });
+    }
+    _fireSoon(...events) {
+        this._bufferedEvents.push(...events);
+        if (this._fireSoonHandle) {
+            clearTimeout(this._fireSoonHandle);
+        }
+        this._fireSoonHandle = setTimeout(() => {
+            this._emitter.fire(this._bufferedEvents);
+            this._bufferedEvents.length = 0;
+        }, 5);
+    }
+}
+exports.OWScriptFS = OWScriptFS;
+//# sourceMappingURL=fileSystem.js.map